==> metaG/R_analyses/Functions/AB_KOmod/07_DA_tab_Kmod.Rmd <==
---
title: "Table KOmod"
output: html_document
editor_options: 
  chunk_output_type: console
---
# Load data
```{r}
library(ggplot2)
library(dplyr)
meta<-readRDS("//metaG//R_analyses/Taxonomy/RDS/meta.rds")
df.cross<-readRDS("/metaG//R_analyses/Functions/AB_KOmod/RDS/df.cross.KOmod.rds")
df.sign<-readRDS("/metaG//R_analyses/Functions/AB_KOmod/RDS/df.sign.KOmod.rds")

meta$PD[meta$PD == "HC"]<-"CTRL"
```

# Format
```{r}

#Single study ODD Ratio 

df.cross.adj<-df.cross %>%
  group_by(Cohort) %>%
  mutate(padjust = p.adjust( p.value, method = "fdr"))

# add dummy variable
df.cross.adj$counts<-ifelse(df.cross.adj$padjust < 0.05 & df.cross.adj$estimate > 0, 1, 0)
df.cross.adj$counts.hc<-ifelse(df.cross.adj$padjust < 0.05 & df.cross.adj$estimate < 0, 1, 0)

# sum for genera
df.cross.adj<- df.cross.adj %>%
  group_by(bin) %>%
  mutate(tot.sign.taxa = sum(counts),
         tot.sign.taxa.hc = sum(counts.hc))

df.sign.uniq<-subset(df.cross.adj[!duplicated(df.cross.adj$bin),], bin %in% df.sign$bins)
df.sign<-df.sign[order(df.sign$TE.random, decreasing = T),]
df.sign.uniq<-df.sign.uniq %>% arrange(factor(bin, levels = df.sign$bins)) 

df.sign.uniq$TE<-vector(length = nrow(df.sign.uniq))
for(i in 1:nrow(df.sign.uniq)){
  df.sign.uniq$TE[i]<-df.sign[which(df.sign$bins == df.sign.uniq$bin[i]),]$TE.random
}
df.sign.uniq$bin<-factor(df.sign.uniq$bin, levels = unique(df.sign.uniq$bin))


# add the ML coefficients
df.pos.neg<-readRDS("/metaG//R_analyses/Functions/AB_KOmod/RDS//Ridge_df.pos.neg.rds")
df.pos.neg<-subset(df.pos.neg, taxa %in% df.sign$bins)

# Format for plotting
df.pos.neg$taxa<-factor(df.pos.neg$taxa, levels = df.sign$bins, ordered = T)

# add dire of enrichment 
df.pos.neg$dir<-vector(length = nrow(df.pos.neg))
for(i in 1:nrow(df.pos.neg)){
  df.pos.neg$dir[i]<-df.sign[df.sign$bins == df.pos.neg$taxa[i],]$dir
}


write.table(df.sign, "/metaG/R_analyses/Functions/AB_KOmod/df.sign.KOmod.txt",
            quote = F,
            col.names = T,
            row.names = F,
            sep = "\t")
write.table(df.pos.neg, "/metaG/R_analyses/Functions/AB_KOmod/df.pos.neg.KOmod.txt",
            quote = F,
            col.names = T,
            row.names = F,
            sep = "\t")
```


# Save a table
```{r}

all(df.sign$bins == df.sign.uniq$bin)

supp.table<-data.frame(KEGG_ID = df.sign$bins,
                       description = df.sign$description, 
                       effect.size = df.sign$TE.random,
                       SE = df.sign$seTE.random,
                       CI.upper = df.sign$upper.random,
                       CI.lower = df.sign$lower.random,
                       direction.enrich = df.sign$dir,
                       p.value.meta = df.sign$pval.random,
                       q.value = df.sign$p.adjust,
                       times.sign.PD = df.sign.uniq$tot.sign.taxa,
                       times.sign.CTRL = df.sign.uniq$tot.sign.taxa.hc,
                       average.ridge.weight = vector(length = nrow(df.sign)),
                       stdev.ridge.weight = vector(length = nrow(df.sign)))

for(i in 1:nrow(supp.table)){
  # add average and stdev
  supp.table$average.ridge.weight[i]<-df.pos.neg[df.pos.neg$taxa == supp.table$KEGG_ID[i],"average"][1,1] %>% unlist() *-1
  supp.table$stdev.ridge.weight[i]<-df.pos.neg[df.pos.neg$taxa == supp.table$KEGG_ID[i],"stdev"][1,1] %>% unlist()
}



head(supp.table)
dim(supp.table)
write.table(x = supp.table, 
            file = "/metaG//R_analyses/Functions/AB_KOmod/RDS/DA.table.complete.txt",
            col.names = T, 
            row.names = F, 
            quote = F,
            sep = "\t")
```
